Make willTransitionTo work
