+++ app/components/InstantMessage/ImMain/Main.jsx
@@ -22,7 +22,77 @@ if (Notification.permission !== 'granted'){
 }
 
 require('./style.less');
+
+function _buildBackEndChannelId(isGroup, channel) {
+  if (isGroup) {
+    return 'team_' + channel.id;
+  } else {
+    var user = LoginStore.getUser();
+    return 'user_' + Math.min(user.id, channel.id) + '_' + Math.max(user.id, channel.id);
+  }
+}
+
+function _init() {
+  var _allTeams = UserStore.getUserInvolvedTeams(LoginStore.getUser().id);
+  var _allUsers = UserStore.getUsersArray();
+  var channels = {
+    publicGroupChannels: _allTeams.map(team=> {
+      var backEndChannelId = _buildBackEndChannelId(true, team);
+      return {
+        isGroup: true,
+        isDirect: false,
+        backEndChannelId: backEndChannelId,
+        id : backEndChannelId,
+        channel : {
+          created_at: team.created_at,
+          name: team.name,
+          updated_at: team.updated_at,
+          id: team.id
+        }
+      }
+    }),
+    directMessageChannels: _allUsers.filter(user => {
+      return '' + user.id !== '' + LoginStore.getUser().id;
+    }).map(user => {
+      var backEndChannelId = _buildBackEndChannelId(false, user);
+      return {
+        isGroup: false,
+        isDirect: true,
+        backEndChannelId: backEndChannelId,
+        id : backEndChannelId,
+        channel : {
+          created_at: user.created_at,
+          name: user.name,
+          realname: user.realname,
+          updated_at: user.updated_at,
+          id: user.id
+        }
+      }
+    })
+  };
+
+  InitAction.init(channels, LoginStore.getUser());
+}
+
 module.exports = React.createClass({
+
+  statics: {
+    willTransitionTo: function (transition, params, query) {
+      let channelIdToGo = params.backEndChannelId;
+      if (channelIdToGo === 'default') {
+        // load from localStorage
+        channelIdToGo = localStorage[IMConstant.LOCALSTORAGE_CHANNEL];
+        transition.redirect('/platform/im/talk/' + channelIdToGo);
+        // redirect
+      } else {
+        if (!SocketStore.getSocket()) {
+          _init();
+        }
+        ChannelAction.changeChannel(channelIdToGo, LoginStore.getUser());
+      }
+    }
+  },
+
   contextTypes: {
     router: React.PropTypes.func.isRequired
   },
@@ -32,6 +102,7 @@ module.exports = React.createClass({
   },
 
   getInitialState() {
+
     return  {
       backEndChannelId : this.context.router.getCurrentParams().backEndChannelId
     }
@@ -39,87 +110,22 @@ module.exports = React.createClass({
 
   componentDidMount() {
     ChannelStore.addChangeListener(this._onChannelChange);
-    if (!SocketStore.getSocket()) {
-      SocketStore.addChangeListener(this._onSocketReady);
-      this._init();
-
-    } else {
-      this._onSocketReady();
-    }
+    //if (!SocketStore.getSocket()) {
+    //  // SocketStore.addChangeListener(this._onSocketReady);
+    //  // this._init();
+    //} else {
+    //  // this._onSocketReady();
+    //}
   },
 
   componentWillUnmount() {
     ChannelStore.removeChangeListener(this._onChannelChange);
-    SocketStore.removeChangeListener(this._onSocketReady);
-  },
-
-  _init() {
-    this._allTeams = UserStore.getUserInvolvedTeams(LoginStore.getUser().id);
-    this._allUsers = UserStore.getUsersArray();
-    var self = this;
-    var channels = {
-      publicGroupChannels: self._allTeams.map(team=> {
-        var backEndChannelId = self._buildBackEndChannelId(true, team);
-          return {
-          isGroup: true,
-          isDirect: false,
-          backEndChannelId: backEndChannelId,
-          id : backEndChannelId,
-          channel : {
-            created_at: team.created_at,
-            name: team.name,
-            updated_at: team.updated_at,
-            id: team.id
-          }
-        }
-      }),
-      directMessageChannels: self._allUsers.filter(user => {
-        return '' + user.id !== '' + LoginStore.getUser().id;
-      }).map(user => {
-        var backEndChannelId = self._buildBackEndChannelId(false, user);
-          return {
-          isGroup: false,
-          isDirect: true,
-          backEndChannelId: backEndChannelId,
-          id : backEndChannelId,
-            channel : {
-              created_at: user.created_at,
-              name: user.name,
-              realname: user.realname,
-              updated_at: user.updated_at,
-              id: user.id
-            }
-        }
-      })
-    };
-
-    InitAction.init(channels, LoginStore.getUser());
+    // SocketStore.removeChangeListener(this._onSocketReady);
   },
 
   _onChannelChange() {
     var currentChannel = ChannelStore.getCurrentChannel();
     this.props.setTitle("Instant Message - Talk - " + (currentChannel.isGroup?currentChannel.channel.name:currentChannel.channel.realname));
-  },
-
-  _onSocketReady() {
-    let channelIdToGo =this.context.router.getCurrentParams().backEndChannelId;
-    if (channelIdToGo === 'default') {
-      // load from localStorage
-      channelIdToGo = localStorage[IMConstant.LOCALSTORAGE_CHANNEL];
-    }
-    if (channelIdToGo) {
-      ChannelAction.changeChannel(channelIdToGo, LoginStore.getUser());
-      this.context.router.transitionTo('/platform/im/talk/' + channelIdToGo);
-    }
-  },
-
-  _buildBackEndChannelId(isGroup, channel) {
-    if (isGroup) {
-      return 'team_' + channel.id;
-    } else {
-      var user = LoginStore.getUser();
-      return 'user_' + Math.min(user.id, channel.id) + '_' + Math.max(user.id, channel.id);
-    }
   },
 
   render() {
@@ -129,7 +135,7 @@ module.exports = React.createClass({
             <ImHistory {...this.props} className="history" ></ImHistory>
             <ImSendBox {...this.props} className="send-box" ></ImSendBox>
           </Flex.Layout>
-          <ImSideNav {...this.props} buildBackEndChannelId={this._buildBackEndChannelId} ></ImSideNav>
+          <ImSideNav {...this.props} buildBackEndChannelId={_buildBackEndChannelId} ></ImSideNav>
         </Flex.Layout>
     );
   }
