+++ app/components/InstantMessage/index.jsx
@@ -3,6 +3,14 @@ const RouteHandler = require("react-router").RouteHandler;
 
 require('./style.less');
 module.exports = React.createClass({
+
+
+  statics: {
+    willTransitionTo: function (transition, params, query, callback) {
+      debugger;
+    }
+  },
+
   componentDidMount() {
 
   },
