+++ package.json
@@ -63,6 +63,7 @@
     "react-hot-loader": "^1.0.0",
     "react-proxy-loader": "^0.3.1",
     "react-router": "^0.13.2",
+    "react-router-proxy-loader": "^0.3.0",
     "react-style": "^0.5.5",
     "react-tap-event-plugin": "^0.1.6",
     "style-loader": "^0.8.2",
